Renamed array_utils to utils. Backwards incompatability is so freeing!
